Fix string conversions for typed values
Because pretty much everything can be converted to strings, specifying a
typed value with desired type "str" basically offered no type validation
against list or other objects instead of that value. What would happen
is that the dict/list would just be converted to a string.

By rejecting any conversion to string that is a dict/list/object in
TypeCastingConverter, using typed values with str will reject any value
that shouldn't be converted to strings.