Implement case-decoupling with case translators
By default, any config specification will be case sensitive: meaning
that if the specification uses snake_case, the config files also need
to use snake_case. Case translators allow the case in the config files
to be decoupled from the case used in the specification (and expected
by the converters: The source case is the case that the config file is
expected to be in. Any specification keys will be translated into that
case before being checked against the input. This allows e.g. json
to use snake_case keys, but yaml to use PascalCase keys. The target case
is the case required by the converters. While this doesn't need to
match the keys in the specification, it is recommended to use that
case, as other settings may lead to confusion (e.g. keys are specified
in snake_case, but converters expect camel_case, may not be completely
understandable when reading it for the first time.