Allow "-1" for array.set and fix type checks
There was an issue where classes as converted type requirements would
be picked up as type checking functions, because they
are technically callable. However, the "call" operator was simply their
__init__ function, which generally does not have the correct
signature for a type check. Now, the signature is verified explicitly.

Additionally, it is now possible to use "[-1]" in overrides to address
the last element in a list (e.g. for appending an object
to a list of unkown size).